fastapi-projects/07-karigar-jobs/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
python-multipart
